% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFinder.R
\name{signature.algorithm}
\alias{signature.algorithm}
\title{Find gene signature or/and built clasifier.}
\usage{
signature.algorithm(table, x, k = 100, m = 20, p = 0.05,
  signature.method = "p.all.med", out = "signature")
}
\arguments{
\item{table}{Data frame with gene in columns and patients or sample
in rows}

\item{x}{Name of target column. (With name ofcancers)}

\item{k}{Number of select genes in first step. (In Kruskal-wallis
statistic). In default k=100.}

\item{m}{Number of important characteristic function in which our
group of gene are enriched. In default m=20.}

\item{p}{Treshold to chose important characteristic function in which our
group of gene are enriched. In default p=0.05.}

\item{signature.method}{Method to calculate the signature. Possible
names: all.all.med, all.min.max, all.5.med, m.all.max, p.all.med.
In default signature.method="p.all.med".}

\item{out}{The parameter describing what we would like to receive
as a result of algorithm. Possible names: signature, SVM.method,
random.forest.method. In default out="signature".}
}
\value{
Signature - vector with genes name. Clasifier of method SVM or
Random Forest.
}
\description{
Find gene signature or/and built clasifier.
}

